(fix) correctly prepend new properties
The previous commit neglected to pass the property cache into useEffect, so each time the socket event handler ran, it ran against the original version of cache, rather than the updated version of cache.

Now, having passed cache into useEffect's watch list, the socket event handler has access to the current state of the cache, and can prepend new items to it.

Also, this commit includes a solution to "add pre-existing properties to
the front of the list". This avoids duplicated properties, and also
slightly reorders the list."